+++ src/mmop-master.lisp
@@ -4,6 +4,7 @@
   (:export pull-master-message
            sent-mmop
            received-mmop
+           worker-task-completed-v0
            ping-v0
            recipe-info-v0
            worker-info-v0
@@ -56,7 +57,9 @@
   ;; client-id type
   (start-node-success-v0 string string)
   ;; client-id type reason-category reason-message
-  (start-node-failure-v0 string string string string))
+  (start-node-failure-v0 string string string string)
+  ;; worker-id node-type
+  (worker-task-completed-v0 string string))
 
 (defmethod create-frames ((message pong-v0))
   `(,(pong-v0%0 message) ,*mmop-v0* "PONG"))
@@ -117,7 +120,9 @@ translate it into an equivalent struct"
                ((list "START-NODE-SUCCESS" node-type)
                 (start-node-success-v0 id node-type))
                ((list "START-NODE-FAILURE" node-type cat msg)
-                (start-node-failure-v0 id node-type cat msg)))))
+                (start-node-failure-v0 id node-type cat msg))
+               ((list "WORKER-TASK-COMPLETED" worker-id node-type)
+                (worker-task-completed-v0 worker-id node-type)))))
 
     (if res res
         (error 'mmop-error :version *mmop-v0* :message "unknown mmop command"))))